Test multiple applications deployed with separate serverInfo config
