+++ dev/io.openliberty.mcp.internal_fat/publish/servers/mcp-server-custom-info/server.xml
@@ -1,4 +1,4 @@
-<server description="MCP Server with Custom ServerInfo">
+<server description="MCP Server with Multiple Apps - Partial and Custom ServerInfo">
 
   <featureManager>
     <feature>servlet-6.0</feature>
@@ -7,8 +7,16 @@
   </featureManager>
  
   
-	<application location="customServerInfoTest.war">
+	<application location="partialServerInfoTest.war">
 	    <mcpServer>
+	        <!-- Only provide name, leave other fields empty/unset to test null handling -->
+	        <info name="partial-server" />
+	    </mcpServer>
+	</application>
+
+	<application location="fullyCustomServerInfoTest.war">
+	    <mcpServer>
+	        <!-- Provide all custom fields to test full configuration -->
 	        <info name="my-custom-server"
 	              title="My Custom MCP Server"
 	              version="2.5.0"
