--- dev/io.openliberty.mcp.internal_fat/publish/servers/mcp-server-partial-info/bootstrap.properties
@@ -1,1 +1,0 @@
-bootstrap.include=../testports.properties
