--- dev/io.openliberty.mcp.internal_fat/publish/servers/mcp-server-partial-info/server.xml
@@ -1,19 +1,0 @@
-<server description="MCP Server with Partial ServerInfo (testing null/empty handling)">
-
-  <featureManager>
-    <feature>servlet-6.0</feature>
-    <feature>cdi-4.0</feature>
-    <feature>mcpServer-1.0</feature>
-  </featureManager>
- 
-  
-	<application location="partialServerInfoTest.war">
-	    <mcpServer>
-	        <!-- Only provide name, leave other fields empty/unset to test null handling -->
-	        <info name="partial-server" />
-	    </mcpServer>
-	</application>
-
-  <include location="../fatTestPorts.xml" />
-  
-</server>
